+++ Assets/01Scripts/InGame/Weapon/Weapon.cs
@@ -2,9 +2,13 @@ using UnityEngine;
 
 public abstract class Weapon : MonoBehaviour
 {
+    [SerializeField] protected WeaponModelView weaponModelView;
+    public InGameManager.StateEnum State { get; protected set; }
     public string WeaponName { get; protected set; }
     public float Atk { get; protected set; }
-    public InGameManager.StateEnum State { get; protected set; } = InGameManager.StateEnum.Running;
+    public Vector3 TargetDir { get; protected set; }
+    [SerializeField] protected float moveSpeed;
+    [SerializeField] protected float deleteTime;
 
     public abstract void Initialize(Collider2D other, float atk); // 정의 해야함
 
@@ -15,19 +19,14 @@ public abstract class Weapon : MonoBehaviour
             other.GetComponent<Enemy>().TakeDamage(Atk);
         }
     }
-    
-    public void OnEnable()
-    {
-        InGameManager.Instance.OnStateChange += OnStateChange;
-    }
 
-    public void OnDisable()
-    {
-        InGameManager.Instance.OnStateChange -= OnStateChange;
-    }
-
-    protected virtual void OnStateChange(InGameManager.StateEnum state)
+    public virtual void HandleOnStateChange(InGameManager.StateEnum state)
     {
         State = state;
+    }
+
+    public virtual void Moving()
+    {
+        transform.position += TargetDir * moveSpeed * Time.deltaTime;
     }
 }
