+++ Assets/01Scripts/InGame/Weapon/WeaponModelView.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 9d9b82f13ffe456f9be2c8110359c6f9
+timeCreated: 1737976679
